// src/data/datasets/homecoming/generated/powersets/corruptor/secondary/kinetics/fulcrum-shift.ts
export const FulcrumShift: Power = {
  "name": "Fulcrum Shift",
  "internalName": "Fulcrum_Shift",
  "available": 29,
  "description": "Fulcrum Shift drains the power of a targeted foe and all foes nearby, transferring it to all adjacent allies, you, and those near you. Affected foes will deal less damage, while your affected allies will deal more. The more foes that are affected, the more power you and your allies receive. Fulcrum Shift can dramatically turn the tide of a battle.Recharge: Slow.",
  "shortHelp": "Ranged (Foe AoE), Foe -DMG, Team +DMG",
  "icon": "kineticboost_kinetictransfer.png",
  "powerType": "Click",
  "targetType": "Foe",
  "effectArea": "SingleTarget",
  "stats": {
    "accuracy": 1,
    "range": 70,
    "recharge": 60,
    "endurance": 15.6,
    "castTime": 2.17
  },
  "allowedEnhancements": [
    "EnduranceReduction",
    "Range",
    "Recharge",
    "Accuracy"
  ],
  "maxSlots": 6,
  "effects": {
    "buffDuration": 45,
    "damageBuff": {
      "scale": 4,
      "table": "Ranged_Buff_Dmg",
      "perTarget": 2
    },
    "damageDebuff": {
      "scale": 2,
      "table": "Ranged_Debuff_Dam"
    },
    "durations": {
      "damageBuff": 45,
      "damageDebuff": 30
    }
  },
  "atoms": [
    ["DamageBuff","Smashing",2,1,45,"Ranged_Buff_Dmg","Str","Magnitude","Target","Any",false,"Stack",2,null,null,1,null,true,null,null,null,null,null,null,2],
    ["DamageBuff","Lethal",2,1,45,"Ranged_Buff_Dmg","Str","Magnitude","Target","Any",false,"Stack",2,null,null,1,null,true,null,null,null,null,null,null,2],
    ["DamageBuff","Fire",2,1,45,"Ranged_Buff_Dmg","Str","Magnitude","Target","Any",false,"Stack",2,null,null,1,null,true,null,null,null,null,null,null,2],
    ["DamageBuff","Cold",2,1,45,"Ranged_Buff_Dmg","Str","Magnitude","Target","Any",false,"Stack",2,null,null,1,null,true,null,null,null,null,null,null,2],
    ["DamageBuff","Energy",2,1,45,"Ranged_Buff_Dmg","Str","Magnitude","Target","Any",false,"Stack",2,null,null,1,null,true,null,null,null,null,null,null,2],
    ["DamageBuff","Negative",2,1,45,"Ranged_Buff_Dmg","Str","Magnitude","Target","Any",false,"Stack",2,null,null,1,null,true,null,null,null,null,null,null,2],
    ["DamageBuff","Psionic",2,1,45,"Ranged_Buff_Dmg","Str","Magnitude","Target","Any",false,"Stack",2,null,null,1,null,true,null,null,null,null,null,null,2],
    ["DamageBuff","Toxic",2,1,45,"Ranged_Buff_Dmg","Str","Magnitude","Target","Any",false,"Stack",2,null,null,1,null,true,null,null,null,null,null,null,2],
    ["DamageBuff","Smashing",2,1,30,"Ranged_Debuff_Dam","Str","Magnitude","Target","Any",true,"Stack",2,null,null,1,null,true],
    ["DamageBuff","Lethal",2,1,30,"Ranged_Debuff_Dam","Str","Magnitude","Target","Any",true,"Stack",2,null,null,1,null,true],
    ["DamageBuff","Fire",2,1,30,"Ranged_Debuff_Dam","Str","Magnitude","Target","Any",true,"Stack",2,null,null,1,null,true],
    ["DamageBuff","Cold",2,1,30,"Ranged_Debuff_Dam","Str","Magnitude","Target","Any",true,"Stack",2,null,null,1,null,true],
    ["DamageBuff","Energy",2,1,30,"Ranged_Debuff_Dam","Str","Magnitude","Target","Any",true,"Stack",2,null,null,1,null,true],
    ["DamageBuff","Negative",2,1,30,"Ranged_Debuff_Dam","Str","Magnitude","Target","Any",true,"Stack",2,null,null,1,null,true],
    ["DamageBuff","Psionic",2,1,30,"Ranged_Debuff_Dam","Str","Magnitude","Target","Any",true,"Stack",2,null,null,1,null,true],
    ["DamageBuff","Toxic",2,1,30,"Ranged_Debuff_Dam","Str","Magnitude","Target","Any",true,"Stack",2,null,null,1,null,true],
    ["DamageBuff","Smashing",4,1,45,"Ranged_Buff_Dmg","Str","Magnitude","Target","Any",false,"Stack",2,null,null,1,null,true],
    ["DamageBuff","Lethal",4,1,45,"Ranged_Buff_Dmg","Str","Magnitude","Target","Any",false,"Stack",2,null,null,1,null,true],
    ["DamageBuff","Fire",4,1,45,"Ranged_Buff_Dmg","Str","Magnitude","Target","Any",false,"Stack",2,null,null,1,null,true],
    ["DamageBuff","Cold",4,1,45,"Ranged_Buff_Dmg","Str","Magnitude","Target","Any",false,"Stack",2,null,null,1,null,true],
    ["DamageBuff","Energy",4,1,45,"Ranged_Buff_Dmg","Str","Magnitude","Target","Any",false,"Stack",2,null,null,1,null,true],
    ["DamageBuff","Negative",4,1,45,"Ranged_Buff_Dmg","Str","Magnitude","Target","Any",false,"Stack",2,null,null,1,null,true],
    ["DamageBuff","Psionic",4,1,45,"Ranged_Buff_Dmg","Str","Magnitude","Target","Any",false,"Stack",2,null,null,1,null,true],
    ["DamageBuff","Toxic",4,1,45,"Ranged_Buff_Dmg","Str","Magnitude","Target","Any",false,"Stack",2,null,null,1,null,true],
    ["ExecutePower",null,0,0,0,"Ranged_Ones","Abs","Magnitude","Target","Any",true,"Stack",2,null,null,1,null,null,null,null,null,null,null,true],
    ["ExecutePower",null,0,0,0,"Ranged_Ones","Abs","Magnitude","Self","Any",true,"Stack",2,null,null,1,null,null,null,null,null,null,null,true]
  ]
};